bionano coverage calculation

+++ src/main/java/cz/vsb/genetics/om/bionano/XMap.java
@@ -15,7 +15,7 @@ public class XMap {
         return entries;
     }
 
-    public Integer getSiteCoverage(Integer contigId, Integer siteId) {
+    public long getSiteCoverage(Integer contigId, Integer siteId) {
         Map<Integer, List<XMapEntry.XMapAlignmentEntry>> contigSites = refSiteToQuerySites.get(contigId);
 
         if (contigSites == null)
